+++ qc-application/src/main/java/com/eager/questioncloud/application/api/creator/CreatorController.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.creator;
 
 import com.eager.questioncloud.application.api.creator.CreatorResponse.CreatorInformationResponse;
-import com.eager.questioncloud.core.domain.creator.CreatorInformation;
+import com.eager.questioncloud.core.domain.creator.dto.CreatorInformation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
