+++ qc-application/src/main/java/com/eager/questioncloud/application/api/creator/CreatorResponse.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.application.api.creator;
 
-import com.eager.questioncloud.core.domain.creator.CreatorInformation;
+import com.eager.questioncloud.core.domain.creator.dto.CreatorInformation;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
