+++ qc-application/src/main/java/com/eager/questioncloud/application/api/creator/CreatorService.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.creator;
 
-import com.eager.questioncloud.core.domain.creator.CreatorInformation;
-import com.eager.questioncloud.core.domain.creator.CreatorRepository;
+import com.eager.questioncloud.core.domain.creator.dto.CreatorInformation;
+import com.eager.questioncloud.core.domain.creator.infrastructure.CreatorRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
