+++ qc-application/src/main/java/com/eager/questioncloud/application/api/creator/RegisterCreatorController.java
@@ -3,8 +3,8 @@ package com.eager.questioncloud.application.api.creator;
 import com.eager.questioncloud.application.api.creator.RegisterCreatorControllerRequest.RegisterCreatorRequest;
 import com.eager.questioncloud.application.api.creator.RegisterCreatorControllerResponse.RegisterCreatorResponse;
 import com.eager.questioncloud.application.security.UserPrincipal;
-import com.eager.questioncloud.core.domain.creator.Creator;
-import com.eager.questioncloud.core.domain.creator.CreatorProfile;
+import com.eager.questioncloud.core.domain.creator.model.Creator;
+import com.eager.questioncloud.core.domain.creator.model.CreatorProfile;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
