+++ qc-application/src/main/java/com/eager/questioncloud/application/api/feed/FeedSubscribeController.java
@@ -3,7 +3,7 @@ package com.eager.questioncloud.application.api.feed;
 import com.eager.questioncloud.application.api.common.PagingResponse;
 import com.eager.questioncloud.application.security.UserPrincipal;
 import com.eager.questioncloud.core.common.PagingInformation;
-import com.eager.questioncloud.core.domain.creator.CreatorInformation;
+import com.eager.questioncloud.core.domain.creator.dto.CreatorInformation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
