+++ qc-application/src/main/java/com/eager/questioncloud/application/api/feed/FeedSubscribeService.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.feed;
 
 import com.eager.questioncloud.core.common.PagingInformation;
-import com.eager.questioncloud.core.domain.creator.CreatorInformation;
+import com.eager.questioncloud.core.domain.creator.dto.CreatorInformation;
 import com.eager.questioncloud.core.domain.subscribe.SubscribeRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
