+++ qc-application/src/main/java/com/eager/questioncloud/application/api/workspace/CreatorProfileService.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.api.workspace;
 
-import com.eager.questioncloud.core.domain.creator.Creator;
-import com.eager.questioncloud.core.domain.creator.CreatorProfile;
-import com.eager.questioncloud.core.domain.creator.CreatorRepository;
+import com.eager.questioncloud.core.domain.creator.infrastructure.CreatorRepository;
+import com.eager.questioncloud.core.domain.creator.model.Creator;
+import com.eager.questioncloud.core.domain.creator.model.CreatorProfile;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
