--- qc-core/src/main/java/com/eager/questioncloud/core/domain/creator/CreatorRepository.java
@@ -1,11 +1,0 @@
-package com.eager.questioncloud.core.domain.creator;
-
-public interface CreatorRepository {
-    Boolean existsById(Long creatorId);
-
-    CreatorInformation getCreatorInformation(Long creatorId);
-
-    Creator save(Creator creator);
-
-    void deleteAllInBatch();
-}
