+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/creator/infrastructure/CreatorEntity.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.creator;
+package com.eager.questioncloud.core.domain.creator.infrastructure;
 
+import com.eager.questioncloud.core.domain.creator.model.Creator;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
