+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/creator/infrastructure/CreatorProfileEntity.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.creator;
+package com.eager.questioncloud.core.domain.creator.infrastructure;
 
+import com.eager.questioncloud.core.domain.creator.model.CreatorProfile;
 import com.eager.questioncloud.core.domain.question.Subject;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.EnumType;
