+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/creator/infrastructure/CreatorStatisticsRepository.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.creator;
+package com.eager.questioncloud.core.domain.creator.infrastructure;
 
+import com.eager.questioncloud.core.domain.creator.model.CreatorStatistics;
 import java.util.List;
 
 public interface CreatorStatisticsRepository {
