+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/QuestionPaymentRepositoryImpl.java
@@ -2,7 +2,7 @@ package com.eager.questioncloud.core.domain.payment;
 
 import static com.eager.questioncloud.core.domain.coupon.infrastructure.QCouponEntity.couponEntity;
 import static com.eager.questioncloud.core.domain.coupon.infrastructure.QUserCouponEntity.userCouponEntity;
-import static com.eager.questioncloud.core.domain.creator.QCreatorEntity.creatorEntity;
+import static com.eager.questioncloud.core.domain.creator.infrastructure.QCreatorEntity.creatorEntity;
 import static com.eager.questioncloud.core.domain.payment.QQuestionPaymentEntity.questionPaymentEntity;
 import static com.eager.questioncloud.core.domain.payment.QQuestionPaymentOrderEntity.questionPaymentOrderEntity;
 import static com.eager.questioncloud.core.domain.question.QQuestionEntity.questionEntity;
