+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/user/UserWithCreator.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.core.domain.user;
 
-import com.eager.questioncloud.core.domain.creator.Creator;
+import com.eager.questioncloud.core.domain.creator.model.Creator;
 
 public record UserWithCreator(User user, Creator creator) {
 }
